chore(executor): extract page + iter + get_or_create modules (#116 step 2)
Three leaf extractions out of executor/mod.rs into their own files:

- page.rs (57 LOC) — `Page<T>` result + `inject_total_count` helper.
- iter.rs (172 LOC) — `ChunkedIter<T>` async iterator (Django's
  `QuerySet.iterator(chunk_size=...)`, issue #23).
- get_or_create.rs (154 LOC) — `get_or_create` / `update_or_create`
  Django-style upsert helpers.

mod.rs: 4830 → 4490 LOC. Net reduction \~340 LOC. Each new file
has no cross-module dependencies beyond `pub use` re-exports from
mod.rs, so the public API is unchanged.

1482 lib tests pass on the sqlite,tenancy litmus.
