+++ crates/rustango/src/sql/executor/mod.rs
@@ -243,57 +243,9 @@ where
     }
 }
 
-/// Result of [`QuerySet::fetch_paginated_on`] — a slice of rows
-/// alongside the total count of matching rows in the underlying
-/// query (i.e. the count *before* LIMIT/OFFSET).
-///
-/// Both pieces come from a single SQL round trip via
-/// `COUNT(*) OVER ()`, so paginated endpoints don't pay the
-/// customary "two queries per page" cost Django's `Paginator`
-/// imposes.
-pub struct Page<T> {
-    pub rows: Vec<T>,
-    pub total: i64,
-}
-
-impl<T> Default for Page<T> {
-    fn default() -> Self {
-        Self {
-            rows: Vec::new(),
-            total: 0,
-        }
-    }
-}
-
-/// Splice `, COUNT(*) OVER () AS "__rustango_total"` into the
-/// compiled SELECT's column list, just before the first ` FROM `.
-/// The Postgres compile_select writer emits the shape
-/// `SELECT <cols> FROM <table> ...` (single space before `FROM`), and
-/// quoted column literals never contain the bare token ` FROM ` —
-/// quoted strings are SqlValue parameters, not part of the column
-/// list. The wrapper-subquery fallback handles unexpected shapes
-/// safely (`COUNT(*) OVER ()` at the OUTER level still counts inner
-/// rows correctly when the inner has no LIMIT, but with LIMIT the
-/// outer COUNT would only see the limited slice — so we depend on
-/// the fast path matching).
-fn inject_total_count(sql: &str) -> String {
-    if let Some(idx) = sql.find(" FROM ") {
-        let mut out = String::with_capacity(sql.len() + 48);
-        out.push_str(&sql[..idx]);
-        out.push_str(", COUNT(*) OVER () AS \"__rustango_total\"");
-        out.push_str(&sql[idx..]);
-        out
-    } else {
-        // Should not reach this branch with the current Postgres
-        // writer — surface the unexpected SQL clearly rather than
-        // silently producing wrong totals.
-        format!(
-            "/* rustango: fetch_paginated_on could not splice COUNT(*) OVER () \
-             into the compiled SELECT — anchor ` FROM ` not found. The query \
-             below will run unchanged but `total` will be 0. */ {sql}"
-        )
-    }
-}
+mod page;
+use page::inject_total_count;
+pub use page::Page;
 
 /// Like [`insert`] but accepts any sqlx executor — `&PgPool`,
 /// `&mut PgConnection`, or a transaction. Tenant-scoped writes need
@@ -4399,146 +4351,8 @@ where
     }
 }
 
-/// v0.45 — Django-style `get_or_create`. Runs the queryset; if it
-/// matches exactly one row return `(row, false)`; if it matches none
-/// invoke `create_fn` to materialize a new instance + insert it and
-/// return `(created, true)`. Matching multiple rows is a
-/// programming error and returns
-/// [`ExecError::MultipleRowsReturned`].
-///
-/// Like Django's helper, this is **not atomic** without an enclosing
-/// transaction — between the SELECT and the INSERT another writer
-/// could insert a colliding row. For race-free behaviour pair it
-/// with `Pool::begin()` or with a UNIQUE constraint that surfaces
-/// the conflict via the existing `upsert()` machinery.
-///
-/// The closure receives an **owned** `Pool` (cheap to clone — it's
-/// an `Arc` internally). That sidesteps Rust's async-closure
-/// lifetime inference: borrowed `&Pool` inside a future returned
-/// from a closure trips the compiler's `'1 must outlive '2` rule.
-///
-/// # Example
-///
-/// ```ignore
-/// let (post, created) = rustango::sql::get_or_create(
-///     Post::objects().filter("slug", "hello"),
-///     |pool| async move {
-///         let mut p = Post {
-///             id: Auto::Unset,
-///             slug: "hello".into(),
-///             title: "Hello".into(),
-///         };
-///         p.insert_pool(&pool).await?;
-///         Ok(p)
-///     },
-///     &pool,
-/// ).await?;
-/// ```
-///
-/// # Errors
-/// - [`ExecError::MultipleRowsReturned`] when the filter matches >1 row.
-/// - Whatever the `create_fn` closure returns when matching no rows.
-/// - Whatever [`FetcherPool::fetch_pool`] returns.
-pub async fn get_or_create<T, F, Fut>(
-    qs: crate::query::QuerySet<T>,
-    create_fn: F,
-    pool: &Pool,
-) -> Result<(T, bool), ExecError>
-where
-    T: Model
-        + MaybePgFromRow
-        + MaybeMyFromRow
-        + MaybeSqliteFromRow
-        + LoadRelated
-        + MaybeMyLoadRelated
-        + MaybeSqliteLoadRelated
-        + Send
-        + Unpin,
-    F: FnOnce(Pool) -> Fut,
-    Fut: std::future::Future<Output = Result<T, ExecError>>,
-{
-    let mut rows = qs.fetch_pool(pool).await?;
-    match rows.len() {
-        0 => Ok((create_fn(pool.clone()).await?, true)),
-        1 => Ok((rows.remove(0), false)),
-        n => Err(ExecError::MultipleRowsReturned {
-            op: "get_or_create",
-            table: T::SCHEMA.table,
-            count: n,
-        }),
-    }
-}
-
-/// v0.45 — Django-style `update_or_create`. Runs the queryset; if
-/// it matches exactly one row, invoke `update_fn` to mutate it +
-/// save the changes and return `(updated, false)`; if it matches
-/// none, invoke `create_fn` and return `(created, true)`. Matching
-/// multiple rows returns [`ExecError::MultipleRowsReturned`].
-///
-/// Same atomicity caveat as [`get_or_create`] — wrap in a
-/// transaction or rely on a UNIQUE constraint for race-free
-/// semantics.
-///
-/// Both closures receive an **owned** `Pool` for the same
-/// async-lifetime reason as [`get_or_create`].
-///
-/// # Example
-///
-/// ```ignore
-/// let (post, created) = rustango::sql::update_or_create(
-///     Post::objects().filter("slug", "hello"),
-///     |pool, mut existing| async move {
-///         existing.title = "New title".into();
-///         existing.save_pool(&pool).await?;
-///         Ok(existing)
-///     },
-///     |pool| async move {
-///         let mut p = Post { /* defaults */ };
-///         p.insert_pool(&pool).await?;
-///         Ok(p)
-///     },
-///     &pool,
-/// ).await?;
-/// ```
-///
-/// # Errors
-/// As [`get_or_create`].
-pub async fn update_or_create<T, UF, UFut, CF, CFut>(
-    qs: crate::query::QuerySet<T>,
-    update_fn: UF,
-    create_fn: CF,
-    pool: &Pool,
-) -> Result<(T, bool), ExecError>
-where
-    T: Model
-        + MaybePgFromRow
-        + MaybeMyFromRow
-        + MaybeSqliteFromRow
-        + LoadRelated
-        + MaybeMyLoadRelated
-        + MaybeSqliteLoadRelated
-        + Send
-        + Unpin,
-    UF: FnOnce(Pool, T) -> UFut,
-    UFut: std::future::Future<Output = Result<T, ExecError>>,
-    CF: FnOnce(Pool) -> CFut,
-    CFut: std::future::Future<Output = Result<T, ExecError>>,
-{
-    let mut rows = qs.fetch_pool(pool).await?;
-    match rows.len() {
-        0 => Ok((create_fn(pool.clone()).await?, true)),
-        1 => {
-            let existing = rows.remove(0);
-            let updated = update_fn(pool.clone(), existing).await?;
-            Ok((updated, false))
-        }
-        n => Err(ExecError::MultipleRowsReturned {
-            op: "update_or_create",
-            table: T::SCHEMA.table,
-            count: n,
-        }),
-    }
-}
+mod get_or_create;
+pub use get_or_create::{get_or_create, update_or_create};
 
 /// v0.45 helper — ensure the queryset has *some* deterministic
 /// ordering before slicing to one row. Used by `first` and
@@ -4616,162 +4430,8 @@ where
     }
 }
 
-/// Chunked async iterator over a compiled query — Django's
-/// `QuerySet.iterator(chunk_size=...)`. Issue #23.
-///
-/// Constructed via [`crate::query::QuerySet::iterator`]. Internally
-/// re-runs the underlying `SELECT` with rotating `OFFSET` per chunk
-/// so the full result set never lives in memory at once. Two
-/// consumption styles:
-///
-/// - [`Self::next_chunk`] yields `Option<Vec<T>>` — a whole batch at
-///   a time. `None` means iteration is done.
-/// - [`Self::next_row`] yields `Option<T>` one at a time, buffering
-///   the current chunk internally between calls.
-///
-/// Mix freely on the same iterator — both methods share the same
-/// buffer + offset state.
-///
-/// Iteration is **exhausted-aware**: once the underlying query
-/// returns a chunk smaller than `chunk_size` (or no rows), the
-/// iterator marks itself exhausted and every subsequent call to
-/// `next_chunk` / `next_row` returns `Ok(None)` without re-querying.
-pub struct ChunkedIter<T> {
-    /// Compiled `SelectQuery` re-issued per chunk with rotating
-    /// `OFFSET`. Cloned every chunk so the per-chunk `limit` /
-    /// `offset` mutation doesn't mutate the original.
-    query: crate::core::SelectQuery,
-    /// Per-chunk row cap. Picked at construction; doesn't change.
-    chunk_size: i64,
-    /// Next `OFFSET` to issue on the next chunk fetch. Bumped by the
-    /// number of rows actually returned (not by `chunk_size`) so a
-    /// short final chunk doesn't skip ahead.
-    offset: i64,
-    /// `true` once a fetched chunk came back smaller than
-    /// `chunk_size`. Further `next_chunk` / `next_row` calls skip the
-    /// DB and return `Ok(None)`.
-    exhausted: bool,
-    /// One-chunk buffer for the row-by-row [`Self::next_row`] path.
-    /// `VecDeque` so `pop_front` is O(1); a `Vec` would be O(n) per
-    /// row. Empty when no rows are buffered or the iterator has been
-    /// drained chunk-by-chunk via [`Self::next_chunk`].
-    buffer: std::collections::VecDeque<T>,
-    /// Cumulative row count yielded so far (across both `next_chunk`
-    /// and `next_row` paths). Used for [`Self::rows_seen`].
-    seen: i64,
-    _model: std::marker::PhantomData<fn() -> T>,
-}
-
-impl<T> ChunkedIter<T>
-where
-    T: Model
-        + MaybePgFromRow
-        + MaybeMyFromRow
-        + MaybeSqliteFromRow
-        + LoadRelated
-        + MaybeMyLoadRelated
-        + MaybeSqliteLoadRelated
-        + Send
-        + Unpin,
-{
-    /// Fetch the next chunk. Returns `Ok(None)` when iteration is done.
-    ///
-    /// If `next_row` was previously called and left rows in the
-    /// internal buffer, those rows are drained as the head of the
-    /// returned chunk before any new DB query — mixing `next_row` and
-    /// `next_chunk` on the same iterator preserves row order.
-    ///
-    /// # Errors
-    /// As [`select_rows_pool_with_related`].
-    pub async fn next_chunk(&mut self, pool: &Pool) -> Result<Option<Vec<T>>, ExecError> {
-        // Drain any rows left in the per-row buffer first — they were
-        // pre-fetched by an earlier `next_row` call and need to come
-        // out before any new DB fetch.
-        if !self.buffer.is_empty() {
-            let buffered: Vec<T> = self.buffer.drain(..).collect();
-            self.seen += buffered.len() as i64;
-            return Ok(Some(buffered));
-        }
-        match self.fetch_next_chunk(pool).await? {
-            Some(rows) => {
-                self.seen += rows.len() as i64;
-                Ok(Some(rows))
-            }
-            None => Ok(None),
-        }
-    }
-
-    /// Yield one row at a time, buffering an internal chunk between
-    /// calls. Returns `Ok(None)` when iteration is done.
-    ///
-    /// O(1) per row — uses `VecDeque::pop_front` against the internal
-    /// buffer, refilling from the next chunk only when empty.
-    ///
-    /// # Errors
-    /// As [`Self::next_chunk`].
-    pub async fn next_row(&mut self, pool: &Pool) -> Result<Option<T>, ExecError> {
-        if let Some(row) = self.buffer.pop_front() {
-            self.seen += 1;
-            return Ok(Some(row));
-        }
-        match self.fetch_next_chunk(pool).await? {
-            Some(rows) => {
-                self.buffer.extend(rows);
-                let row = self.buffer.pop_front();
-                if row.is_some() {
-                    self.seen += 1;
-                }
-                Ok(row)
-            }
-            None => Ok(None),
-        }
-    }
-
-    /// Shared fetch path for [`Self::next_chunk`] / [`Self::next_row`].
-    /// Clones the compiled query, rotates `LIMIT`/`OFFSET` for this
-    /// chunk, runs the fetch, advances `offset`, and flips the
-    /// `exhausted` flag when the chunk comes back short or empty.
-    /// Does NOT touch `self.seen` — that's the caller's
-    /// responsibility because the row-by-row path counts as rows pop
-    /// out of the buffer, not as the chunk arrives.
-    async fn fetch_next_chunk(&mut self, pool: &Pool) -> Result<Option<Vec<T>>, ExecError> {
-        if self.exhausted {
-            return Ok(None);
-        }
-        // Clone so the per-chunk LIMIT/OFFSET tweaks don't mutate the
-        // original (the iterator may be re-used across many chunks).
-        let mut q = self.query.clone();
-        q.limit = Some(self.chunk_size);
-        q.offset = Some(self.offset);
-        let rows = select_rows_pool_with_related::<T>(pool, &q).await?;
-        let n = rows.len() as i64;
-        if rows.is_empty() {
-            self.exhausted = true;
-            return Ok(None);
-        }
-        if n < self.chunk_size {
-            self.exhausted = true;
-        }
-        self.offset += n;
-        Ok(Some(rows))
-    }
-
-    /// Cumulative count of rows yielded so far across both
-    /// `next_chunk` and `next_row` calls. Useful for progress
-    /// reporting on long drains.
-    #[must_use]
-    pub fn rows_seen(&self) -> i64 {
-        self.seen
-    }
-
-    /// `true` once iteration has been exhausted — every subsequent
-    /// `next_chunk` / `next_row` call will return `Ok(None)` without
-    /// hitting the database.
-    #[must_use]
-    pub fn is_exhausted(&self) -> bool {
-        self.exhausted && self.buffer.is_empty()
-    }
-}
+mod iter;
+pub use iter::ChunkedIter;
 
 #[cfg(test)]
 mod pool_dispatch_tests {
